Mark zeroconf.cache as protected by renaming to zeroconf._cache
- The public API should only access zeroconf and zeroconf.aio
  as internals may be relocated between releases
